ml-service/requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
python-multipart
pillow
numpy
scikit-learn
matplotlib
opencv-python
seaborn
